feat: post detail component
